+++ src/structures/diagnostic.php
@@ -6,8 +6,10 @@ enum ErrorType: string {
     case Lexical = "Lexical";
     case Syntax = "Syntax";
     case Runtime = "Runtime";
-    case FloatingPoint = "Floating Point";
-    case UndeclaredIdentifier = "Undeclared Identifier";
+    case Import = "Import";
+    case FloatingPoint = "FloatingPoint";
+    case InvalidArgumentCount = "InvalidArgumentCount";
+    case Identifier = "Identifier";
 
 
     public function str() {
@@ -23,7 +25,6 @@ class Diagnostic {
     public $position;
 
     public function raise($error_type, $message, $start, $cursor) {
-
         error_log("File \"" . $cursor->filepath . ":" . $cursor->line . ":" . $cursor->column . "\", line " . $cursor->line . PHP_EOL);
         error_log($this->display_snippet($start, $cursor));
 
@@ -34,7 +35,6 @@ class Diagnostic {
 
     public function display_snippet($start, $cursor) {
 
-        $start = $start;
         $end = $cursor->pos;
         $lines = explode(PHP_EOL, substr($cursor->source, 0, $start));
 
@@ -42,4 +42,3 @@ class Diagnostic {
     }
 }
 
-?>