+++ src/structures/nodes.php
@@ -14,7 +14,6 @@ class EchoStatement {
     }
 }
 
-
 class DeclareVariable {
     public $name;
     public $value;
@@ -31,28 +30,35 @@ class AugAssign {
     public $name;
     public $operator; 
     public $value;
+    public $pos;
 
-    public function __construct($name, $operator, $value) {
+    public function __construct($name, $operator, $value, $pos) {
         $this->name = $name;
         $this->value = $value;
         $this->operator = $operator;
+        $this->pos = $pos;
     }
 }
 
 class DeleteVariable {
     public $name;
+    public $pos;
 
-    public function __construct($name) {
+    public function __construct($name, $pos) {
         $this->name = $name;
+        $this->pos = $pos;
     }
 }
 
 class ObjectDeclare {
     public $name;
     public $body = [];
-    public function __construct($name, $body) {
+    public $pos;
+
+    public function __construct($name, $body, $pos) {
         $this->name = $name;
         $this->body = $body;
+        $this->pos = $pos;
     }
 }
 
@@ -61,35 +67,41 @@ class DeclareFunction {
     public $args;
     public $body;
     public $inclass;
+    public $pos;
 
-    public function __construct($name, $args, $body, $inclass = FALSE) {
+    public function __construct($name, $args, $body, $pos, $inclass = FALSE) {
         $this->name = $name;
         $this->args = $args;
         $this->body = $body;
         $this->inclass = $inclass;
+        $this->pos = $pos;
     }
 }
 
 class MemberAccess {
     public $object;
     public $property;
-    
-    public function __construct($object, $property) {
+    public $pos;
+
+    public function __construct($object, $property, $pos) {
         $this->object = $object;
         $this->property = $property;
+        $this->pos = $pos;
     }
 }
-
 
 class FunctionCall {
     public $name;
     public $args;
+    public $pos;
 
-    public function __construct($name, $args) {
+    public function __construct($name, $args, $pos) {
         $this->name = $name;
         $this->args = $args;
+        $this->pos = $pos;
     }
 }
+
 class IfStatement {
     public $condition;
     public $body = [];
@@ -108,66 +120,107 @@ class IfStatement {
 
 class NumberLiteral {
     public $value;
+    public $pos;
 
-    public function __construct($value) {
+    public function __construct($value, $pos) {
         $this->value = $value;
+        $this->pos = $pos;
     }
 }
 
 class StringLiteral {
     public $value;
+    public $pos;
 
-    public function __construct($value) {
+    public function __construct($value, $pos) {
         $this->value = $value;
+        $this->pos = $pos;
     }
 }
 
 class Identifier {
     public $value;
+    public $pos;
 
-    public function __construct($value) {
+    public function __construct($value, $pos) {
         $this->value = $value;
+        $this->pos = $pos;
     }
 }
 
 class IncludeStatement {
     public $filepath;
+    public $pos;
 
-    public function __construct($filepath) {
+    public function __construct($filepath, $pos) {
         $this->filepath = $filepath;
+        $this->pos = $pos;
     }
 }
 
 class ImportStatement {
     public $filepath;
-    #public $alias;
-    public function __construct($filepath) {
+    public $pos;
+
+    public function __construct($filepath, $pos) {
         $this->filepath = $filepath;
-        #$this->alias = $alias;
+        $this->pos = $pos;
     }
 }
 
 class ReturnStatement {
     public $value;
+    public $pos;
 
-    public function __construct($value) {
+    public function __construct($value, $pos) {
         $this->value = $value;
+        $this->pos = $pos;
     }
 }
-class Kill {}
-class None {}
-class BreakLoop {}
-class ContinueLoop {}
+
+class Kill {
+    public $pos;
+
+    public function __construct($pos) {
+        $this->pos = $pos;
+    }
+}
+
+class None {
+    public $pos;
+
+    public function __construct($pos) {
+        $this->pos = $pos;
+    }
+}
+
+class BreakLoop {
+    public $pos;
+
+    public function __construct($pos) {
+        $this->pos = $pos;
+    }
+}
+
+class ContinueLoop {
+    public $pos;
+
+    public function __construct($pos) {
+        $this->pos = $pos;
+    }
+}
 
 class BinaryOperation {
     public $right;
     public $operator;
     public $left;
+    public $pos;
 
-    public function __construct($right, $operator, $left) {
+    public function __construct($right, $operator, $left, $pos) {
         $this->right = $right;
         $this->operator = $operator;
         $this->left = $left;
+        $this->pos = $pos;
     }
 }
 
@@ -175,21 +228,25 @@ class BoolOp {
     public $right;
     public $operator;
     public $left;
+    public $pos;
 
-    public function __construct($right, $operator, $left) {
+    public function __construct($right, $operator, $left, $pos) {
         $this->right = $right;
         $this->operator = $operator;
         $this->left = $left;
+        $this->pos = $pos;
     }
 }
 
 class Subscript {
     public $identifier;
     public $slice;
+    public $pos;
 
-    public function __construct($value, $slice) {
+    public function __construct($value, $slice, $pos) {
         $this->identifier = $value;
         $this->slice = $slice;
+        $this->pos = $pos;
     }
 }
 
@@ -197,19 +254,23 @@ class Slice {
     public $lower;
     public $upper;
     public $steps;
+    public $pos;
 
-    public function __construct($lower=null, $upper=null, $steps=null) {
+    public function __construct($lower=null, $upper=null, $steps=null, $pos) {
         $this->lower = $lower;
         $this->upper = $upper;
         $this->steps = $steps;
+        $this->pos = $pos;
     }
 }
 
 class Index {
     public $value;
+    public $pos;
 
-    public function __construct($value) {
+    public function __construct($value, $pos) {
         $this->value = $value;
+        $this->pos = $pos;
     }
 }
 
@@ -229,6 +290,7 @@ enum Unary
     case Minus;
     case Negate;
 }
+
 enum Operation
 {
     case Add;
@@ -250,21 +312,25 @@ class Compare {
     public $right;
     public $operator;
     public $left;
+    public $pos;
 
-    public function __construct($right, $operator, $left) {
+    public function __construct($right, $operator, $left, $pos) {
         $this->right = $right;
         $this->operator = $operator;
         $this->left = $left;
+        $this->pos = $pos;
     }
 }
 
 class UnaryOperation {
     public $operator;
     public $operand;
+    public $pos;
 
-    public function __construct($operator, $operand) {
+    public function __construct($operator, $operand, $pos) {
         $this->operator = $operator;
-        $this->left = $operand;
+        $this->operand = $operand;
+        $this->pos = $pos;
     }
 }
 
@@ -272,6 +338,7 @@ class LoopStatement {
     public $condition;
     public $body;
     public $pos;
+
     public function __construct($condition, $body, $pos) {
         $this->condition = $condition;
         $this->body = $body;
@@ -282,30 +349,38 @@ class LoopStatement {
 class AssignVariable {
     public $left;
     public $value;
-    public function __construct($name, $value) {
+    public $pos;
+
+    public function __construct($name, $value, $pos) {
         $this->left = $name;
         $this->value = $value;
-    }
-}
-class LiteralText {
-    public $value;
-    public function __construct($value) {
-        $this->value = $value;
+        $this->pos = $pos;
     }
 }
 
+class LiteralText {
+    public $value;
+    public $pos;
+
+    public function __construct($value, $pos) {
+        $this->value = $value;
+        $this->pos = $pos;
+    }
+}
 
 class FunctionObject {
     public $name;
     public $args;
     public $scope;
     public $body;
+    public $pos;
 
-    public function __construct($name, $args, $body, $scope) {
+    public function __construct($name, $args, $body, $scope, $pos) {
         $this->name = $name;
-        $this->args= $args;
+        $this->args = $args;
         $this->body = $body;
         $this->scope = $scope;
+        $this->pos = $pos;
     }
 }
 
@@ -313,12 +388,13 @@ class ObjectType {
     public $name;
     public $properties;
     public $interpreter;
+    public $pos;
 
-    public function __construct($name, $properties, $interpreter = null) {
+    public function __construct($name, $properties, $pos, $interpreter = null) {
         $this->name = $name;
         $this->properties = $properties;
         $this->interpreter = $interpreter;
+        $this->pos = $pos;
     }
 }
 
-?>