+++ CHANGELOG.md
@@ -18,4 +18,4 @@
 - Introduced inline comment syntax `#` to allow for commenting out lines of code.
 - Added `exit` with the ability to print a message and exit the program.
 - Added script-loader for module-like system in source code.
-- Added Error System for Lumen.
+- Added Error & Cursor System for Lumen.